requirements.txt
# Web Frameworks
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP/Networking
requests==2.31.0
httpx==0.25.2

# Data Processing & Visualization
pandas==2.1.4
plotly==5.18.0

# Data Validation
pydantic==2.5.0

# AI/OpenAI
openai==1.3.5

# Environment Configuration
python-dotenv==1.0.0

# FastAPI Dependencies
python-multipart==0.0.6
aiofiles==23.2.1

# Type Hints
typing-extensions==4.8.0

# HuggingFace / Vision
torch==2.0.1
diffusers==0.21.0
transformers==4.33.0
accelerate==0.22.0
safetensors==0.3.3
Pillow==10.0.0
gradio_client

# UI
gradio==4.16.0


spaces
torch==2.1.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
Pillow==10.1.0
